added local DynamoDB test
